api admin

+++ tc100/src/main/java/com/java/dao/ActiveDao.java
@@ -1,5 +1,6 @@
 package com.java.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.java.base.BaseDao;
@@ -13,7 +14,7 @@ public interface ActiveDao extends BaseDao<Active>{
 
 	int addViewCount(String actId);
 
-	void operateAct(String actId, int amount);
+	void operateAct(String actId, BigDecimal amount);
 
 	int getActCount(String userId, String actType, String state);
